Database normalization (from arrays to tables) Sessions addition for tokens

+++ src/main/kotlin/org/penakelex/database/extenstions/LocalDate.kt
@@ -7,11 +7,11 @@ import java.time.LocalDate
  * @receiver [LocalDate] date of birth of the person
  * @return [Int] - age of the person
  * */
-fun LocalDate.getAgeOfPersonToday(): Int {
+fun LocalDate.getAgeOfPersonToday(): UShort {
     val todayDate = LocalDate.now()
     return if (todayDate.monthValue - this.monthValue > 0
         || todayDate.monthValue - this.monthValue == 0
         && todayDate.dayOfMonth - this.dayOfMonth >= 0
-    ) todayDate.year - this.year
-    else todayDate.year - this.year - 1
+    ) (todayDate.year - this.year).toUShort()
+    else (todayDate.year - this.year - 1).toUShort()
 }
